@@ -15,14 +15,14 @@
 	},
 	"require": {
 		"ext-openssl": "*",
-		"php": "^5.6.0|^7.0"
+		"php": "^7.0"
 	},
 	"require-dev": {
 		"consistence/coding-standard": "^0.10.0",
 		"jakub-onderka/php-parallel-lint": "^0.9",
 		"phing/phing": "~2.12",
-		"phpunit/phpunit": "^5.0.4",
-		"slevomat/coding-standard": "^1.0.5",
+		"phpunit/phpunit": "^5.2.12",
+		"slevomat/coding-standard": "dev-php7#5dceec0",
 		"guzzlehttp/guzzle": "^6.1",
 		"satooshi/php-coveralls": "^1.0"
 	},
