@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types = 1);
 
 namespace SlevomatCsobGateway\Api\Driver;
 
@@ -7,9 +7,9 @@ use SlevomatCsobGateway\Api\ApiClientDriverException;
 class GuzzleDriverException extends \RuntimeException implements ApiClientDriverException
 {
 
-	public function __construct(\Exception $e)
+	public function __construct(\Throwable $previous)
 	{
-		parent::__construct('Request error: ' . $e->getMessage(), $e->getCode(), $e);
+		parent::__construct('Request error: ' . $previous->getMessage(), $previous->getCode(), $previous);
 	}
 
 }
