@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types = 1);
 
 namespace SlevomatCsobGateway\Api;
 
@@ -10,21 +10,13 @@ class InvalidPaymentException extends RequestException
 	/** @var ProcessPaymentRequest */
 	private $request;
 
-	/**
-	 * @param ProcessPaymentRequest $request
-	 * @param Response $response
-	 * @param string $payId
-	 */
-	public function __construct(ProcessPaymentRequest $request, Response $response, $payId)
+	public function __construct(ProcessPaymentRequest $request, Response $response, string $payId)
 	{
 		parent::__construct(sprintf('PayId %s is invalid or expired.', $payId), $response);
 		$this->request = $request;
 	}
 
-	/**
-	 * @return ProcessPaymentRequest
-	 */
-	public function getRequest()
+	public function getRequest(): ProcessPaymentRequest
 	{
 		return $this->request;
 	}
