@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types = 1);
 
 namespace SlevomatCsobGateway\Crypto;
 
@@ -8,29 +8,17 @@ class SignatureDataFormatter
 	/** @var mixed[] */
 	private $keysPriority;
 
-	/**
-	 * @param mixed[] $keysPriority
-	 */
 	public function __construct(array $keysPriority)
 	{
 		$this->keysPriority = $keysPriority;
 	}
 
-	/**
-	 * @param mixed[] $data
-	 * @return string
-	 */
-	public function formatDataForSignature(array $data)
+	public function formatDataForSignature(array $data): string
 	{
 		return implode('|', $this->generateMessage($data, $this->keysPriority));
 	}
 
-	/**
-	 * @param mixed[] $data
-	 * @param mixed[] $keys
-	 * @return mixed[]
-	 */
-	private function generateMessage(array $data, array $keys)
+	private function generateMessage(array $data, array $keys): array
 	{
 		$message = [];
 
@@ -45,9 +33,11 @@ class SignatureDataFormatter
 				}
 
 			} else {
-				$message[] = is_bool($data[$key])
-					? ($data[$key] ? 'true' : 'false')
-					: $data[$key];
+				if (is_bool($data[$key])) {
+					$message[] = $data[$key] ? 'true' : 'false';
+				} else {
+					$message[] = $data[$key];
+				}
 			}
 		}
 
