@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types = 1);
 
 namespace SlevomatCsobGateway\Crypto;
 
@@ -10,9 +10,6 @@ class SigningFailedException extends \RuntimeException
 	 */
 	private $data;
 
-	/**
-	 * @param mixed[] $data
-	 */
 	public function __construct(array $data)
 	{
 		parent::__construct('Signing failed');
@@ -20,10 +17,7 @@ class SigningFailedException extends \RuntimeException
 		$this->data = $data;
 	}
 
-	/**
-	 * @return mixed[]
-	 */
-	public function getData()
+	public function getData(): array
 	{
 		return $this->data;
 	}
