@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types = 1);
 
 namespace SlevomatCsobGateway\Type;
 
@@ -24,7 +24,7 @@ class InvalidEnumValueException extends \InvalidArgumentException
 		parent::__construct(sprintf(
 			'Invalid enum value \'%s\'. Available values: %s',
 			$value,
-			join(', ', $availableValues)
+			implode(', ', $availableValues)
 		));
 
 		$this->value = $value;
@@ -32,7 +32,7 @@ class InvalidEnumValueException extends \InvalidArgumentException
 	}
 
 	/**
-	 * @return string
+	 * @return mixed
 	 */
 	public function getValue()
 	{
@@ -42,7 +42,7 @@ class InvalidEnumValueException extends \InvalidArgumentException
 	/**
 	 * @return mixed[]
 	 */
-	public function getAvailableValues()
+	public function getAvailableValues(): array
 	{
 		return $this->availableValues;
 	}
