@@ -1,3 +1,3 @@
-<?php
+<?php declare(strict_types = 1);
 
 require __DIR__ . '/../vendor/autoload.php';
